# Conflicts:
#	ses-api/ses-api-common/src/main/java/com/redescooter/ses/api/common/constant/JedisConstant.java
#	ses-app/ses-mobile-client/src/main/resources/config/application-prod.yml
#	ses-app/ses-mobile-rps/src/main/resources/config/application-prod.yml
#	ses-app/ses-web-delivery/src/main/resources/config/application-prod.yml
#	ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/service/sys/impl/StaffServiceImpl.java
#	ses-app/ses-web-ros/src/main/resources/config/application-local.yml
#	ses-app/ses-web-ros/src/main/resources/config/application-prod.yml
#	ses-service/ses-service-base/ses-service-foundation/src/main/resources/config/application-prod.yml
#	ses-service/ses-service-base/ses-service-proxy/src/main/resources/config/application-prod.yml
#	ses-service/ses-service-base/ses-service-scooter/src/main/resources/config/application-prod.yml
#	ses-service/ses-service-hub/src/main/resources/config/application-prod.yml
#	ses-service/ses-service-mobile-b/src/main/resources/config/application-prod.yml
#	ses-service/ses-service-mobile-c/src/main/resources/config/application-prod.yml

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/service/restproduction/impl/SaleCombinServiceImpl.java
@@ -6,6 +6,8 @@ import com.redescooter.ses.api.common.constant.JedisConstant;
 import com.redescooter.ses.api.common.vo.base.GeneralResult;
 import com.redescooter.ses.api.common.vo.base.IdEnter;
 import com.redescooter.ses.api.common.vo.base.PageResult;
+import com.redescooter.ses.api.hub.service.website.PartsService;
+import com.redescooter.ses.api.hub.vo.website.SyncSalePartsDataEnter;
 import com.redescooter.ses.starter.common.service.IdAppService;
 import com.redescooter.ses.starter.redis.service.JedisService;
 import com.redescooter.ses.tool.utils.SesStringUtils;
@@ -24,12 +26,13 @@ import com.redescooter.ses.web.ros.vo.restproduct.CombinNameEnter;
 import com.redescooter.ses.web.ros.vo.restproduct.SaleCombinListResult;
 import com.redescooter.ses.web.ros.vo.restproduct.SaleCombinSaveOrUpdateEnter;
 import com.redescooter.ses.web.ros.vo.restproduct.SaleListEnter;
+import io.seata.spring.annotation.GlobalTransactional;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.dubbo.config.annotation.DubboReference;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.List;
@@ -60,8 +63,11 @@ public class SaleCombinServiceImpl implements SaleCombinService {
     @DubboReference
     private IdAppService idAppService;
 
+    @DubboReference
+    private PartsService partsService;
+
     @Override
-    @Transactional(rollbackFor = Exception.class)
+    @GlobalTransactional(rollbackFor = Exception.class)
     public GeneralResult saveSaleCombin(SaleCombinSaveOrUpdateEnter enter) {
         // 去空格
         enter = SesStringUtils.objStringTrim(enter);
@@ -108,7 +114,7 @@ public class SaleCombinServiceImpl implements SaleCombinService {
     }
 
     @Override
-    @Transactional(rollbackFor = Exception.class)
+    @GlobalTransactional(rollbackFor = Exception.class)  
     public GeneralResult editSaleCombin(SaleCombinSaveOrUpdateEnter enter) {
         // 去空格
         enter = SesStringUtils.objStringTrim(enter);
@@ -129,7 +135,13 @@ public class SaleCombinServiceImpl implements SaleCombinService {
 
     @Override
     public GeneralResult deleteSaleCombin(IdEnter enter) {
+        OpeSaleCombin combin = opeSaleCombinService.getById(enter.getId());
+        if (null == combin) {
+            throw new SesWebRosException(ExceptionCodeEnums.PRODUCT_IS_NOT_EXIST.getCode(), ExceptionCodeEnums.PRODUCT_IS_NOT_EXIST.getMessage());
+        }
         opeSaleCombinService.removeById(enter.getId());
+        // ros删除配件后,需将官网已同步的配件同样删除
+        syncDeleteData(combin.getProductCode());
         return new GeneralResult(enter.getRequestId());
     }
 
@@ -149,7 +161,31 @@ public class SaleCombinServiceImpl implements SaleCombinService {
         Integer saleStatus = combin.getSaleStutas();
         combin.setSaleStutas(saleStatus == 0 ? 1 : 0);
         opeSaleCombinService.updateById(combin);
+
+        // 数据同步到官网的销售配件
+        OpeSaleCombin opeSaleCombin = opeSaleCombinService.getById(enter.getId());
+        syncData(opeSaleCombin);
         return new GeneralResult(enter.getRequestId());
+    }
+
+    @Async
+    void syncData(OpeSaleCombin combin) {
+        SyncSalePartsDataEnter model = new SyncSalePartsDataEnter();
+        model.setStatus(combin.getSaleStutas() == 1 ? 1 : 2);
+        model.setPartsType(2);
+        model.setPartsNumber(combin.getProductName());
+        model.setEnName(combin.getCombinName());
+        model.setEffectiveTime(new Date());
+        model.setRemark(combin.getRemark());
+        model.setRevision(0);
+        model.setCreatedBy(0L);
+        model.setCreatedTime(new Date());
+        partsService.syncSalePartsData(model);
+    }
+
+    @Async
+    void syncDeleteData(String productCode) {
+        partsService.syncDeleteData(productCode);
     }
